TravisCI: Let Galaxy use venv
If I recall correctly, since TravisCI has already put
Python in a venv and we're only using this as single
use VM, it was faster to re-use that venv.

That worked fine for my galaxy_mira and pico_galaxy
repositories, but something in the job splitting
when testing with parallelism enabled breaks when
trying to find the Python module six.

So, let's just let Galaxy do its usual thing with
virtual environments - this is closer to a standard
installation after all.